feat: split Settings into League and Account homes
