+++ apps/web/e2e/prod/screenshots.spec.ts
@@ -21,8 +21,9 @@ const ROUTES: Array<{ slug: string; path: string }> = [
   { slug: "seasons", path: "/dashboard/seasons" },
   { slug: "divisions", path: "/dashboard/teams?view=divisions" },
   { slug: "discover", path: "/dashboard/discover" },
-  { slug: "import", path: "/dashboard/import" },
-  { slug: "billing", path: "/dashboard/billing" },
+  { slug: "settings", path: "/dashboard/settings" },
+  { slug: "import", path: "/dashboard/settings/account/import" },
+  { slug: "billing", path: "/dashboard/settings/account/billing" },
   { slug: "roles", path: "/dashboard/roles" },
   { slug: "legal-privacy", path: "/privacy" },
   { slug: "legal-terms", path: "/terms" },
