+++ apps/web/src/app/dashboard/import/page.tsx
@@ -1,16 +1,10 @@
-import { ImportForm } from "./_components/import-form";
-import { NflSyncCard } from "./_components/nfl-sync-card";
+import { permanentRedirect } from "next/navigation";
+import { accountImportHref } from "@/components/workspace/resource-navigation";
 
-export default function ImportPage() {
-  return (
-    <div>
-      <h2 className="mb-6 text-lg font-semibold text-foreground">
-        Import Data
-      </h2>
-      <div className="space-y-8">
-        <NflSyncCard />
-        <ImportForm />
-      </div>
-    </div>
-  );
+/**
+ * Legacy Import URL. The Import UI moved under Account Settings (issue #576,
+ * ASR-8); this path only remains as a permanent redirect source.
+ */
+export default function LegacyImportRedirect() {
+  permanentRedirect(accountImportHref());
 }
