+++ apps/web/src/lib/stripe-checkout.ts
@@ -34,8 +34,8 @@ export async function createSubscriptionCheckoutUrl(
     mode: "subscription",
     customer: customerId,
     line_items: [{ price: priceId, quantity: 1 }],
-    success_url: `${appUrl}/dashboard/billing?success=true&session_id={CHECKOUT_SESSION_ID}`,
-    cancel_url: `${appUrl}/dashboard/billing?cancelled=true`,
+    success_url: `${appUrl}/dashboard/settings/account/billing?success=true&session_id={CHECKOUT_SESSION_ID}`,
+    cancel_url: `${appUrl}/dashboard/settings/account/billing?cancelled=true`,
     client_reference_id: userId,
     metadata: { clerkUserId: userId },
     subscription_data: { metadata: { clerkUserId: userId } },
