+++ docs/archive/issue-574-progress.txt
@@ -11,6 +11,6 @@ On completion, move this file and the plan to docs/archive/.
     [x] 3.1 - Unit regression for no ?from= emitters
     [x] 3.2 - Playwright roster → Player Home URL (search empty)
     [x] 3.3 - type-check + focused e2e green (team-detail + players + data-table: 25/25)
-[ ] 4.0 - Ship & archive
-    [ ] 4.1 - Open PR with Closes #574
-    [ ] 4.2 - On merge: move docs/issue-574-* to docs/archive/
+[x] 4.0 - Ship & archive
+    [x] 4.1 - Open PR with Closes #574 (merged as #587)
+    [x] 4.2 - Archived docs/issue-574-* to docs/archive/
